fix: rename packages to gitsemver
