+++ main.go
@@ -1,10 +1,10 @@
 package main
 
 import (
-	git_semver "github.com/carlsberg/git-semver/cmd/git-semver"
+	"github.com/carlsberg/git-semver/cmd/gitsemver"
 	"github.com/spf13/cobra"
 )
 
 func main() {
-	cobra.CheckErr(git_semver.Execute())
+	cobra.CheckErr(gitsemver.Execute())
 }
